+++ web/utils.js
@@ -43,19 +43,55 @@ export async function	updateAccessTkn()
 	}
 }
 
-export function	clearErrFields()
+export function	clearPopovers()
 {
-	let errMsg = document.getElementById("errMsg");
-	if (errMsg !== null)
-		errMsg.remove();
 	const	markedFields = document.querySelectorAll(".is-invalid");
 	markedFields.forEach((field) => {
 		field.classList.remove("is-invalid");
 	} );
 
-	let errPopover = document.querySelectorAll(".popover");
-	errPopover.forEach(msg => {
+	let popoverMsgs = document.querySelectorAll(".popover");
+	popoverMsgs.forEach(msg => {
 		msg.remove();
 	})
 		
+}
+
+
+export function showPopover(msg, targetElem, color = 'primary')
+{
+		const popover = new bootstrap.Popover(targetElem);
+
+		// Show the popover on page load
+
+
+		const observer = new MutationObserver(() => {
+			const popoverEl = document.querySelector('.popover');
+
+			if (popoverEl) {
+				popoverEl.classList.add('border', `border-${color}`);
+
+				const body = popoverEl.querySelector('.popover-body');
+				if (body) {
+					body.classList.add(`text-${color}`);
+					body.innerText = msg
+				}
+
+				observer.disconnect(); // Stop watching once found
+			}
+		});
+
+		observer.observe(document.body, { childList: true, subtree: true });
+
+		popover.show();
+		setTimeout(() => {
+			popover.hide();
+			setTimeout(() => {
+				popover.dispose();
+			}, 200);
+			const	markedFields = document.querySelectorAll(".is-invalid");
+			markedFields.forEach((field) => {
+			field.classList.remove("is-invalid");
+	} );
+		}, 3000);
 }
