+++ web/views/account_settings.html
@@ -22,20 +22,20 @@
 				<div class="tab-pane fade show active" id="changeUsername">
 					<div class="col-7 col-lg-3 mx-auto">
 						<form id="changeUsernameForm">
-							<div class="input-group my-2">
-								<span class="input-group-text" data-bs-placement="right" data-bs-content=" ">
+							<div class="input-group my-2" data-bs-placement="right" data-bs-content=" ">
+								<span class="input-group-text">
 									<span class="bi-person-fill"></span>
 								</span>
 								<input id="newUserField" type="text" class="form-control" placeholder="New Username">
 							</div>
-							<button type="submit" class="btn btn-outline-light mt-2">Confirm</button>
+							<button id="chgUsernameBtn" type="submit" class="btn btn-outline-light mt-2">Confirm</button>
 						</form>
 					</div>
 				</div>
 				<!--CHANGE MOTTO-->
 				<div class="tab-pane fade" id="changeMotto">
 					<div class="col-7 mx-auto">
-						<form id="changeMottoForm">
+						<form id="changeMottoForm" data-bs-placement="right" data-bs-content=" ">
 							<div class="input-group my-2">
 								<textarea id="newMottoField" type="text" class="form-control" placeholder="New Motto"></textarea>
 							</div>
@@ -48,7 +48,7 @@
 					<div class="col-7 col-lg-3 gap-2 mx-auto">
 						<form id="changePfpForm">
 							<div class="my-2">
-								<input id="pfpUploadBtn" type="file" class="btn btn-light" accept="image/*"></input>
+								<input data-bs-placement="right" data-bs-content=" " id="pfpUploadBtn" type="file" class="btn btn-light" accept="image/*"></input>
 							</div>
 							<button type="submit" class="btn btn-outline-light mt-2">Confirm</button>
 						</form>
