+++ web/views/auth.html
@@ -33,7 +33,7 @@
 									</span>
 									<input id="loginPasswordField" type="password" class="form-control" placeholder="Password">
 								</div>
-								<button type="submit" class="btn btn-primary mt-2">Login</button>
+								<button id="loginBtn" type="submit" class="btn btn-primary mt-2">Login</button>
 							</form>
 						</div>
 					</div>
